requirements_cpanel.txt
# Core Flask dependencies for cPanel hosting
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
SQLAlchemy==2.0.41
Werkzeug==2.3.7
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
alembic==1.16.2
Mako==1.3.10
python-dateutil==2.9.0.post0
email_validator==2.2.0
python-dotenv==1.1.0
